Extract BikeCodec from network-manager (#318 Step 2)
First increment of the @usersfirst/room split: separate the game-specific
binary message schema (bike physics state/lean/pedal/event/profile) from
the generic transport, so the transport sends/receives opaque bytes and
knows nothing about bikes.

- New js/net/bike-codec.js: a pure BikeCodec owning the wire schema and the
  pre-allocated 60Hz STATE/LEAN send buffers (encodeState/decodeState,
  encodeLean/decodeLean, encodePedal/decodePedal, encodeEvent,
  encodeProfile/decodeProfile).
- network-manager.js delegates all encode/decode to the codec; HEARTBEAT
  (generic keepalive) and the MSG-type dispatch stay in the transport for
  now (that glue moves in a later increment).

Pure refactor — wire format is byte-for-byte unchanged. Verified with a
round-trip test: STATE=46B exact, -1 timer sentinel, legacy 42B (timer-less)
still decodes, lean/pedal/event/profile all round-trip.
